Refactor: Place 상세 위젯 분리 #99
